feat: 소켓 차량 패키지 추가 #40
