+++ src/main/java/com/douzone/server/config/socket/ChatConfig.java
@@ -2,19 +2,22 @@ package com.douzone.server.config.socket;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.web.socket.WebSocketHandler;
 import org.springframework.web.socket.config.annotation.EnableWebSocket;
 import org.springframework.web.socket.config.annotation.WebSocketConfigurer;
 import org.springframework.web.socket.config.annotation.WebSocketHandlerRegistry;
 
-
 @RequiredArgsConstructor
 @Configuration
 @EnableWebSocket
-public class WebSocketConfig implements WebSocketConfigurer {
-	private final WebSocketTestHandler webSocketTestHandler;
+public class ChatConfig implements WebSocketConfigurer {
+
+	private final WebSocketHandler webSocketHandler;
 
 	@Override
 	public void registerWebSocketHandlers(WebSocketHandlerRegistry registry) {
-		registry.addHandler(webSocketTestHandler, "/ws/room").setAllowedOrigins("*");
+		registry.addHandler(webSocketHandler, "ws/chat").setAllowedOrigins("*");
 	}
+
 }
+
