+++ src/main/java/com/douzone/server/config/socket/ChatService.java
@@ -1,4 +1,4 @@
-package com.douzone.server.config.socket2;
+package com.douzone.server.config.socket;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.RequiredArgsConstructor;
@@ -42,7 +42,7 @@ public class ChatService {
 	}
 
 	public <T> void sendMessage(WebSocketSession session, T message) {
-		try{
+		try {
 			session.sendMessage(new TextMessage(objectMapper.writeValueAsString(message)));
 		} catch (IOException e) {
 			log.error(e.getMessage(), e);
