+++ src/main/java/com/douzone/server/config/socket/vehicle/VehicleSocketController.java
@@ -1,0 +1,4 @@
+package com.douzone.server.config.socket.vehicle;
+
+public class VehicleSocketController {
+}
