+++ src/main/java/com/douzone/server/config/socket/vehicle/VehicleSocketDTO.java
@@ -1,0 +1,4 @@
+package com.douzone.server.config.socket.vehicle;
+
+public class VehicleSocketDTO {
+}
