--- src/main/java/com/douzone/server/config/socket2/ChatConfig.java
@@ -1,22 +1,0 @@
-package com.douzone.server.config.socket2;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.messaging.simp.config.MessageBrokerRegistry;
-import org.springframework.web.socket.WebSocketHandler;
-import org.springframework.web.socket.config.annotation.*;
-
-@RequiredArgsConstructor
-@Configuration
-@EnableWebSocket
-public class ChatConfig implements WebSocketConfigurer {
-
-	private final WebSocketHandler webSocketHandler;
-
-	@Override
-	public void registerWebSocketHandlers(WebSocketHandlerRegistry registry) {
-		registry.addHandler(webSocketHandler, "ws/chat").setAllowedOrigins("*");
-	}
-
-}
-
